backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
crewai==0.80.0
langchain-openai==0.2.0
openai>=1.0.0,<2.0.0
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.0.0
supabase>=2.3.0
huggingface_hub>=0.23.0
playwright>=1.47.0
python-dateutil>=2.8.2
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0

